Allow INIT_HEAP_SIZE to be set at compile time
I was having problems with the init_heap being too small, as shown by the error message:
malloc_count ### init heap full !!!

The solution is to increase the heap size. Before this change, that could only be done by editing the source code.
With this change the init_heap size can be controlled by setting the define with the compiler.